@@ -1,6 +1,6 @@
 module github.com/regen-network/regen-ledger/api/v2
 
-go 1.18
+go 1.19
 
 require (
 	github.com/cosmos/cosmos-proto v1.0.0-beta.1
